requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
et_xmlfile==2.0.0
Flask==3.1.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-documentai==3.5.0
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.73.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
rsa==4.9.1
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
